Server/requirements.txt
asgiref==3.9.1
Django==5.2.4
djangorestframework==3.16.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
django-environ==0.11.0

psycopg2-binary==2.9.9          # kết nối PostgreSQL
django-cors-headers==4.3.1      # dùng trong middleware
djangorestframework-simplejwt==5.3.1  # cho JWT
django-stubs[compatible]>=4.2.4
djangorestframework-stubs[compatible]>=3.14
django-filter>=23.3 # cho filter trong DRF sửa lại requirements nhé
channels==4.0.0
channels-redis==4.1.0
redis==5.0.0